feat(assistant-context): move `updateAssistantContext` into a service
